Move private-types to types/private
